[SWM-144] refactor: corrtect typo "timeTraking" -> "timeTracking"
