+++ src/main/java/com/climbx/climbx/common/timeTracking/TimeTrackingInterceptor.java
@@ -1,4 +1,4 @@
-package com.climbx.climbx.common.timeTraking;
+package com.climbx.climbx.common.timeTracking;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
